[BUGFIX][t3editor] (De-)Activating does not work in Chrome
The checkbox "Deactivate t3editor" does not work in modern Chrome
versions (e.g. 18). It works in Firefox, but didn't even work while
integrating the code into TYPO3 in Chrome.

Prototype recognizes it as Webkit and thus does not disable/enable it,
which was a pretty weird implementation, which is now removed.

Change-Id: I9660af7c754d25b1321c329e9d67f80ba6d71db8
Fixes: #33515
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/8976
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Simon Schaufelberger
Tested-by: Simon Schaufelberger
Reviewed-by: Steffen Gebert
Tested-by: Steffen Gebert