[TASK] Remove internal sys_history cleanup feature
The internal cleanup feature for the table sys_history
kept the entries only for 30 days by default. This behaviour
can lead to lost information if the history is needed and
the feature wasn't known and changed.

As a replacement, the users can use the table garbage
collection scheduler task that exists for some time now.

Change-Id: Iede559e0ece282c16a54e6659c26737c8bf86d90
Resolves: #38436
Releases: 6.0
Reviewed-on: http://review.typo3.org/12404
Reviewed-by: Markus Klein
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Wouter Wolters
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert