[!!!][TASK] Add unique identifier to language files
Add an identifier based on timestamp to all language files. This
number must not be changed if a file is moved around. The numbers
are unique throughout the project. From now on, each language file
requires this number.

This change is required for the translation server, see blueprint
for details: http://wiki.typo3.org/Blueprints/TranslationHandling.

Resolves: #62892
Releases: master, 6.2
Change-Id: If4b7483f68860fd144ee12c83cc5f1d468c610ce
Reviewed-on: http://review.typo3.org/34048
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>