[CLEANUP] ClickMenu internal code cleanup
As a followup to the API changes
of the clickmenu within #62709 this
change removes any leftovers
of the internals of the clickmenu
using e.g. non-sprite icons, and
removing all non-AJAX-functionality
which was used via alt_clickmenu.php
without &ajax=1 (for the very old topbar
in pre-4.2).

Additionally, it deprecates alt_clickmenu.php,
which was already announced in #62709. The
prehistoric user UC option "noMenuMode"
was also removed completely.

A final followup will bring a non-table output
as well as cleaned up JS calls, and a JSON
based data exchange.

Relates: #62709
Relates: #18742
Resolves: #62594
Resolves: #63103
Change-Id: I6867f559c90a32f7f8f588bb9ea7a754d7602260
Reviewed-on: http://review.typo3.org/34399
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>