[CLEANUP] Add missing empty lines after copyright comments I
At the moment in some classes you can find an empty line between the
copyright comment and the first PHPdoc bock comment. In some other
classes there is no such empty line.

This commit adds an empty line between those comments if there is no
empty line.

This is not part of the TYPO3 CMS CGL but increases consistency in this
area.

This regular expression was used to add the empty lines:

Search:
( \* The TYPO3 project - inspiring people to share!\n \*/\n)(/\*\*\n \*)

Replace:
$1\n$2

Resolves: #63828
Releases: master
Change-Id: I1d921921a3f60a62c29de3300b25b5df36a43b0a
Reviewed-on: http://review.typo3.org/35390
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>