[FEATURE] Introduce .typoscript file extension
The new file extension .typoscript will be the default for TypoScript
configuration files and is the only recommended one from now on. This
effort is made to introduce a dedicated file extension for TypoScript
configuration files, and to avoid conflicts with already existing and
more spread file extensions like ".ts" for TypeScript or Video Transport
Stream Files.

Resolves: #78161
Resolves: #80689
Releases: master, 8.7
Change-Id: I44b41631f498e3ba39e0f1936329094a59859f4e
Reviewed-on: https://review.typo3.org/50096
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>