[TASK] Streamline ext_localconf.php and ext_tables.php
The goal is to minimize TYPO3_MODE usage in ext_localconf and ext_tables and
provide a clear process for extension developers where to put what.

This extension streamlines the whole usage within all system extensions,
and documents the expected behaviour of all third-party extensions, which
will also be included in docs.typo3.org.

Resolves: #82692
Releases: master
Change-Id: Id83ff0dfc6198564443f2f42e273bcd4f6f25e3e
Reviewed-on: https://review.typo3.org/52533
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>