[!!!][TASK] Drop dedicated edit view in EXT:tstemplate
Triggering ext:t3editor out-of-bound of FormEngine is pretty tricky and
blocks further development of t3editor. To streamline this, the
current implementation of t3editor within ext:tstemplate is revamped
and now uses FormEngine directly to edit "constants" and "setup" of
template records.
Some hooks are now obsolete and dropped along the way.

Resolves: #81171
Releases: master
Change-Id: Iab62a15828bfb82bc9f8ad11f705bb4529a38734
Reviewed-on: https://review.typo3.org/52696
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>