[!!!][FEATURE] Refactor EXT:t3editor
EXT:t3editor is refactored to use the latest CodeMirror version (at the
time of writing it's 5.28.0).
Developers are now able to easily register and use custom modes for syntax
highlighting. Also, addons for CodeMirror can be added easily.

As CodeMirror provides these already, some useful addons are already
registered, for example a search with hit annotations, fullscreen mode and
jump-to-line functionality. A positive side-effect is that a lot of custom
code was removed.

Since the architecture allows to register additional modes, the TypoScript
mode has been refactored and might be moved into a separate extension in
further iterations.

Resolves: #81901
Releases: master
Change-Id: I432d1fdb3ba1f2a0b9246207dd70843b4bdbfbd9
Reviewed-on: https://review.typo3.org/53413
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>