[TASK] Update codemirror to 5.30 and drop codemirror-minified again
The package `codemirror-minified` used as a drop-in replacement of
`codemirror` is not maintained well and is replaced with the original
package again. To still get the benefits of page loading speed, at least
the JavaScript files are compressed by our own toolchain now.

CSS can't get compressed now as the original copyright gets replaced by
our own "banner".

Resolves: #82713
Releases: master
Change-Id: Ic12a7e9e70c63f2c0ae322e470453fe056c99155
Reviewed-on: https://review.typo3.org/54346
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>