[BUGFIX] Prevent closing a record after T3Editor content changes
This patch prevents closing a form without showing the modal
to inform the backend user about unsaved changes if values of
an element with renderType `t3editor` were changed within the
CodeMirror editor.

In addition to the fix, a CSS border is added if the code editor
content was changed to improve the user experience.

Releases: master
Resolves: #86380
Change-Id: I641180a531b879ece70e30d0a22c9e5878d02953
Reviewed-on: https://review.typo3.org/58394
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>