[TASK] Remove leftover "menu_type" occurrences
With TYPO3 v8.6, the new Content Types for menus were introduced
where each menu type is not stored in the database field "menu_type"
anymore, but moved to a separate "CType" property. This change
was breaking, however some left-over code was there but never
removed. The code does not do anything anymore but handle the
special "menu_type" cases like in Page Module CType Previews
or categorization handling.

FYI: This is a pre-patch to streamline the category API.

Resolves: #90035
Releases: master
Change-Id: I180fc5b1e2d33f74cf05d2565ab11049cc361c49
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62785
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benni Mack <benni@typo3.org>