[TASK] Set php header comment via php-cs-fixer
With this patch, the header comment of php files is
automatically added by the php-cs-fixer, which guarantees
that its format and place of occurrence remain the same
in all files.

Files that are copied over from other projects are excluded.

Furthermore, files that are kind of inspired by other
projects also get the same header comment but may have
a second, additional comment explaining its origin.

Used command:

    bin/php-cs-fixer fix --config=Build/php-cs-fixer/header-comment.php

Releases: master
Resolves: #91024
Change-Id: I5a040517e0fbde6e5a27d589bf2f222078326dc8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64159
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>