[!!!][TASK] Remove built-in support for accesskeys in HMENU
The HTML standard "accesskey" is built into TYPO3's
HMENU functionality, with only limited support:

* No cross-platform labeling possible
* No concrete definition possible via TypoScript (AccessPad)
* Random access based on the menu items' title
* Legacy implementation based on TSFE state

In addition, proper browser support has been poorly,
and it is only recommended for specific use-cases,
not really for a11y purposes.

For this reason, this feature is removed as is, and
should be implemented in the TYPO3 Frontend when it is
needed in custom extensions or via JavaScript support:

https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/accesskey#accessibility_concerns
https://webaim.org/standards/wcag/checklist#:~:text=accesskey%20should%20typically%20be%20avoided

The HMENU property "accessKey" will stop working
in TYPO3 v12.0.

Side-note: An additional, leftover reference ("collapse") in tsref.xml
was also removed for HMENU.

Resolves: #96708
Releases: main
Change-Id: I27a86f582d7c2f009465cb465bdb1a0de392a4df
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73216
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>