[TASK] Fix npm copy of contrib packages
@lezer/common – in contrast to other @lezer modules,
is placed in dist/index.js instead of dist/index.es.js.
Therefore we add a distinct copy rule.

The sortablejs npm package recently changed to provide both,
a UMD bundle and a native ESM module.
As we actually failed to update the UMD bundle (due to a
rename of the build output filename in sortablejs),
we switch to the new ESM variant right away.

muuri failed to update because dashboardToEs6
was missing in the npmcopy concurrent target.

Resolves: #101784
Resolves: #101785
Resolves: #101787
Related: #101783
Releases: main, 12.4
Change-Id: I2ccf69847876edb55f089a4a5142ac388699fb54
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80687
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>